breaking-bad/requirements.txt
pymongo[srv]
google-generativeai
python-dotenv
